small file restructure/rename - rename files to more terse names to avoid redundancy - move actual pages to their proper location
