+++ client/src/components/Customer/Dashboard/Editor/Utils/Verdict.tsx
